removed parseInt from client and replaced it with parse float

+++ client/src/components/multiStepDivs/carDivs/CarComponets/carModalSchema.ts
@@ -68,10 +68,10 @@ export const carModalSchema = z
       .string({
         required_error: "Please enter a price",
       })
-      .refine((item) => parseInt(item) >= 1000, {
+      .refine((item) => parseFloat(item) >= 1000, {
         message: "Please enter a price greater than or equal to $1,000",
       })
-      .refine((item) => parseInt(item) <= 750000, {
+      .refine((item) => parseFloat(item) <= 750000, {
         message: "Please enter a price less than or equal to $750,000",
       }),
     mileage: z.string({
@@ -98,7 +98,7 @@ export const carModalSchema = z
       .string({
         required_error: "Please enter your salary",
       })
-      .refine((item) => parseInt(item) <= 1000000, {
+      .refine((item) => parseFloat(item) <= 1000000, {
         message: "Please enter an income/salary less than 1,000,000",
       }),
       term: z.number({
@@ -114,7 +114,7 @@ export const carModalSchema = z
     img: z.any().optional(),
   })
   .superRefine((values, ctx) => {
-    if (parseInt(values.downPayment) >= parseInt(values.price)) {
+    if (parseFloat(values.downPayment) >= parseFloat(values.price)) {
       ctx.addIssue({
         message: "Your down payment should not be greater the car price.",
         code: z.ZodIssueCode.custom,
